+++ dsfilter/R3/utils.py
@@ -110,7 +110,7 @@ def bilinear_interpolate(
     v01: ti.f32,
     v10: ti.f32,
     v11: ti.f32,
-    r: ti.types.vector(3, ti.i32)
+    r: ti.types.vector(3, ti.f32)
 ) -> ti.f32:
     """
     @taichi.func
